Fixing broken mapbox token paths after refactoring.
