+++ src/applications/facility-locator/containers/VAMap.jsx
@@ -6,7 +6,8 @@ import { browserHistory } from 'react-router';
 import { connect } from 'react-redux';
 import { Tabs, TabList, TabPanel, Tab } from 'react-tabs';
 import { Map, TileLayer, FeatureGroup } from 'react-leaflet';
-import { mapboxClient, mapboxToken } from '../components/MapboxClient';
+import { mapboxClient } from '../components/MapboxClient';
+import { mapboxToken } from '../utils/mapboxToken';
 import isMobile from 'ismobilejs';
 import { isEmpty, debounce } from 'lodash';
 import appendQuery from 'append-query';
